fixed make warnings

+++ src/mem.c
@@ -17,34 +17,34 @@ void memset(uint8 *dest, uint8 val, uint32 len)
 void print_mem_values()
 {
   vga_write("Start: ");
-  vga_put_hex(base);
+  vga_put_hex((uint32)base);
   vga_put('\n');
 
   vga_write("Start of text: ");
-  vga_put_hex(text);
+  vga_put_hex((uint32)text);
   vga_put('\n');
 
   vga_write("End of text: ");
-  vga_put_hex(etext);
+  vga_put_hex((uint32)etext);
   vga_put('\n');
 
   vga_write("Start of data: ");
-  vga_put_hex(data);
+  vga_put_hex((uint32)data);
   vga_put('\n');
 
   vga_write("End of data: ");
-  vga_put_hex(edata);
+  vga_put_hex((uint32)edata);
   vga_put('\n');
 
   vga_write("Start of bss: ");
-  vga_put_hex(bss);
+  vga_put_hex((uint32)bss);
   vga_put('\n');
 
   vga_write("End of bss: ");
-  vga_put_hex(ebss);
+  vga_put_hex((uint32)ebss);
   vga_put('\n');
 
   vga_write("End: ");
-  vga_put_hex(end);
+  vga_put_hex((uint32)end);
   vga_put('\n');
 }
